+++ kotlinbard/src/main/kotlin/io/github/enjoydambience/kotlinbard/Access.java
@@ -1,0 +1,11 @@
+package io.github.enjoydambience.kotlinbard;
+
+import com.squareup.kotlinpoet.PropertySpec;
+import com.squareup.kotlinpoet.TypeName;
+
+class Access {
+    static TypeName getTypeOf(PropertySpec.Builder builder) {
+        //noinspection KotlinInternalInJava
+        return builder.getType$kotlinpoet();
+    }
+}