+++ kotlinbard/src/test/kotlin/io/github/enjoydambience/kotlinbard/ReadmeExamples.kt
@@ -58,7 +58,7 @@ class ReadmeExamples : StringSpec({
             get {
                 addStatement("return field")
             }
-            set("value", String::class) {
+            set("value") {
                 addStatement("field = value")
             }
             //or, for parameterless set